backend/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
vaderSentiment==3.3.2
supabase==2.15.2
lxml==4.9.3
python-dotenv==1.0.0 